Refactor: re-organized api calls
